ScheduleDailyCheck update
