+++ petition_service/main.go
@@ -39,8 +39,6 @@ func grpcStart(petitionSvc rpc.IPetitionService) {
 
 	pb.RegisterPetitionServiceServer(s, server)
 
-	rpc.ScheduleDailyCheck(server)
-
 	slog.Infof("gRPC Server listening at %v\n", lis.Addr())
 
 	if err := s.Serve(lis); err != nil {
